kernel apply mips patch

+++ core_repos/src/libs/3_8_0/linux_kernel_2.6.22.19/include/asm-mips/cpu-info.h
@@ -41,8 +41,8 @@ struct cache_desc {
 #define MIPS_CACHE_PINDEX	0x00000020	/* Physically indexed cache */
 
 struct cpuinfo_mips {
-	unsigned long		udelay_val;
-	unsigned long		asid_cache;
+	unsigned int		udelay_val;
+	unsigned int		asid_cache;
 #if defined(CONFIG_SGI_IP27)
 //	cpuid_t		p_cpuid;	/* PROM assigned cpuid */
 	cnodeid_t	p_nodeid;	/* my node ID in compact-id-space */
